+++ rand_distr/src/unit_ball.rs
@@ -31,7 +31,7 @@ pub struct UnitBall;
 impl<F: Float + SampleUniform> Distribution<[F; 3]> for UnitBall {
     #[inline]
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> [F; 3] {
-        let uniform = Uniform::new(F::from(-1.).unwrap(), F::from(1.).unwrap());
+        let uniform = Uniform::new(F::from(-1.).unwrap(), F::from(1.).unwrap()).unwrap();
         let mut x1;
         let mut x2;
         let mut x3;