+++ src/distributions/uniform.rs
@@ -10,9 +10,8 @@
 //! A distribution uniformly sampling numbers within a given range.
 //!
 //! [`Uniform`] is the standard distribution to sample uniformly from a range;
-//! e.g. `Uniform::new_inclusive(1, 6)` can sample integers from 1 to 6, like a
-//! standard die. [`Rng::gen_range`] supports any type supported by
-//! [`Uniform`].
+//! e.g. `Uniform::new_inclusive(1, 6).unwrap()` can sample integers from 1 to 6, like a
+//! standard die. [`Rng::gen_range`] supports any type supported by [`Uniform`].
 //!
 //! This distribution is provided with support for several primitive types
 //! (all integer and floating-point types) as well as [`std::time::Duration`],
@@ -31,7 +30,7 @@
 //! use rand::distributions::Uniform;
 //!
 //! let mut rng = thread_rng();
-//! let side = Uniform::new(-10.0, 10.0);
+//! let side = Uniform::new(-10.0, 10.0).unwrap();
 //!
 //! // sample between 1 and 10 points
 //! for _ in 0..rng.gen_range(1..=10) {
@@ -49,11 +48,11 @@
 //!
 //! At a minimum, the back-end needs to store any parameters needed for sampling
 //! (e.g. the target range) and implement `new`, `new_inclusive` and `sample`.
-//! Those methods should include an assert to check the range is valid (i.e.
+//! Those methods should include an assertion to check the range is valid (i.e.
 //! `low < high`). The example below merely wraps another back-end.
 //!
 //! The `new`, `new_inclusive` and `sample_single` functions use arguments of
-//! type SampleBorrow<X> in order to support passing in values by reference or
+//! type SampleBorrow<X> to support passing in values by reference or
 //! by value. In the implementation of these functions, you can choose to
 //! simply use the reference returned by [`SampleBorrow::borrow`], or you can choose
 //! to copy or clone the value, whatever is appropriate for your type.
@@ -61,7 +60,7 @@
 //! ```
 //! use rand::prelude::*;
 //! use rand::distributions::uniform::{Uniform, SampleUniform,
-//!         UniformSampler, UniformFloat, SampleBorrow};
+//!         UniformSampler, UniformFloat, SampleBorrow, Error};
 //!
 //! struct MyF32(f32);
 //!
@@ -70,20 +69,18 @@
 //!
 //! impl UniformSampler for UniformMyF32 {
 //!     type X = MyF32;
-//!     fn new<B1, B2>(low: B1, high: B2) -> Self
+//!
+//!     fn new<B1, B2>(low: B1, high: B2) -> Result<Self, Error>
 //!         where B1: SampleBorrow<Self::X> + Sized,
 //!               B2: SampleBorrow<Self::X> + Sized
 //!     {
-//!         UniformMyF32(UniformFloat::<f32>::new(low.borrow().0, high.borrow().0))
+//!         UniformFloat::<f32>::new(low.borrow().0, high.borrow().0).map(UniformMyF32)
 //!     }
-//!     fn new_inclusive<B1, B2>(low: B1, high: B2) -> Self
+//!     fn new_inclusive<B1, B2>(low: B1, high: B2) -> Result<Self, Error>
 //!         where B1: SampleBorrow<Self::X> + Sized,
 //!               B2: SampleBorrow<Self::X> + Sized
 //!     {
-//!         UniformMyF32(UniformFloat::<f32>::new_inclusive(
-//!             low.borrow().0,
-//!             high.borrow().0,
-//!         ))
+//!         UniformFloat::<f32>::new_inclusive(low.borrow().0, high.borrow().0).map(UniformMyF32)
 //!     }
 //!     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Self::X {
 //!         MyF32(self.0.sample(rng))
@@ -95,7 +92,7 @@
 //! }
 //!
 //! let (low, high) = (MyF32(17.0f32), MyF32(22.0f32));
-//! let uniform = Uniform::new(low, high);
+//! let uniform = Uniform::new(low, high).unwrap();
 //! let x = uniform.sample(&mut thread_rng());
 //! ```
 //!
@@ -106,8 +103,10 @@
 //! [`UniformDuration`]: crate::distributions::uniform::UniformDuration
 //! [`SampleBorrow::borrow`]: crate::distributions::uniform::SampleBorrow::borrow
 
+use core::fmt;
 use core::time::Duration;
 use core::ops::{Range, RangeInclusive};
+use std::convert::TryFrom;
 
 use crate::distributions::float::IntoFloat;
 use crate::distributions::utils::{BoolAsSIMD, FloatAsSIMD, FloatSIMDUtils, WideningMultiply};
@@ -119,6 +118,29 @@ use crate::{Rng, RngCore};
 use crate::distributions::utils::Float;
 
 #[cfg(feature = "simd_support")] use packed_simd::*;
+
+/// Error type returned from `Uniform::new{,_inclusive}`.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum Error {
+    /// `low > high`, or equal in case of exclusive range.
+    EmptyRange,
+    /// Input or range `high - low` is non-finite. Not relevant to integer types. In release mode,
+    /// only the range is checked.
+    NonFinite,
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(match self {
+            Error::EmptyRange => "low > high (or equal if exclusive) in uniform distribution",
+            Error::NonFinite => "Non-finite range in uniform distribution",
+        })
+    }
+}
+
+#[cfg(feature = "std")]
+#[cfg_attr(doc_cfg, doc(cfg(feature = "std")))]
+impl std::error::Error for Error {}
 
 #[cfg(feature = "serde1")]
 use serde::{Serialize, Deserialize};
@@ -132,10 +154,10 @@ use serde::{Serialize, Deserialize};
 ///
 /// When sampling from a constant range, many calculations can happen at
 /// compile-time and all methods should be fast; for floating-point ranges and
-/// the full range of integer types this should have comparable performance to
+/// the full range of integer types, this should have comparable performance to
 /// the `Standard` distribution.
 ///
-/// Steps are taken to avoid bias which might be present in naive
+/// Steps are taken to avoid bias, which might be present in naive
 /// implementations; for example `rng.gen::<u8>() % 170` samples from the range
 /// `[0, 169]` but is twice as likely to select numbers less than 85 than other
 /// values. Further, the implementations here give more weight to the high-bits
@@ -149,9 +171,10 @@ use serde::{Serialize, Deserialize};
 /// # Example
 ///
 /// ```
+/// use std::convert::TryFrom;
 /// use rand::distributions::{Distribution, Uniform};
 ///
-/// let between = Uniform::from(10..10000);
+/// let between = Uniform::try_from(10..10000).unwrap();
 /// let mut rng = rand::thread_rng();
 /// let mut sum = 0;
 /// for _ in 0..1000 {
@@ -179,24 +202,24 @@ use serde::{Serialize, Deserialize};
 pub struct Uniform<X: SampleUniform>(X::Sampler);
 
 impl<X: SampleUniform> Uniform<X> {
-    /// Create a new `Uniform` instance which samples uniformly from the half
-    /// open range `[low, high)` (excluding `high`). Panics if `low >= high`.
-    pub fn new<B1, B2>(low: B1, high: B2) -> Uniform<X>
+    /// Create a new `Uniform` instance, which samples uniformly from the half
+    /// open range `[low, high)` (excluding `high`). Fails if `low >= high`.
+    pub fn new<B1, B2>(low: B1, high: B2) -> Result<Uniform<X>, Error>
     where
         B1: SampleBorrow<X> + Sized,
         B2: SampleBorrow<X> + Sized,
     {
-        Uniform(X::Sampler::new(low, high))
+        X::Sampler::new(low, high).map(Uniform)
     }
 
-    /// Create a new `Uniform` instance which samples uniformly from the closed
-    /// range `[low, high]` (inclusive). Panics if `low > high`.
-    pub fn new_inclusive<B1, B2>(low: B1, high: B2) -> Uniform<X>
+    /// Create a new `Uniform` instance, which samples uniformly from the closed
+    /// range `[low, high]` (inclusive). Fails if `low > high`.
+    pub fn new_inclusive<B1, B2>(low: B1, high: B2) -> Result<Uniform<X>, Error>
     where
         B1: SampleBorrow<X> + Sized,
         B2: SampleBorrow<X> + Sized,
     {
-        Uniform(X::Sampler::new_inclusive(low, high))
+        X::Sampler::new_inclusive(low, high).map(Uniform)
     }
 }
 
@@ -232,22 +255,18 @@ pub trait UniformSampler: Sized {
     /// The type sampled by this implementation.
     type X;
 
-    /// Construct self, with inclusive lower bound and exclusive upper bound
-    /// `[low, high)`.
+    /// Construct self, with inclusive lower bound and exclusive upper bound `[low, high)`.
     ///
-    /// Usually users should not call this directly but instead use
-    /// `Uniform::new`, which asserts that `low < high` before calling this.
-    fn new<B1, B2>(low: B1, high: B2) -> Self
+    /// Usually users should not call this directly but instead use `Uniform::new`.
+    fn new<B1, B2>(low: B1, high: B2) -> Result<Self, Error>
     where
         B1: SampleBorrow<Self::X> + Sized,
         B2: SampleBorrow<Self::X> + Sized;
 
     /// Construct self, with inclusive bounds `[low, high]`.
     ///
-    /// Usually users should not call this directly but instead use
-    /// `Uniform::new_inclusive`, which asserts that `low <= high` before
-    /// calling this.
-    fn new_inclusive<B1, B2>(low: B1, high: B2) -> Self
+    /// Usually users should not call this directly but instead use `Uniform::new_inclusive`.
+    fn new_inclusive<B1, B2>(low: B1, high: B2) -> Result<Self, Error>
     where
         B1: SampleBorrow<Self::X> + Sized,
         B2: SampleBorrow<Self::X> + Sized;
@@ -279,7 +298,7 @@ pub trait UniformSampler: Sized {
         B1: SampleBorrow<Self::X> + Sized,
         B2: SampleBorrow<Self::X> + Sized,
     {
-        let uniform: Self = UniformSampler::new(low, high);
+        let uniform: Self = UniformSampler::new(low, high).unwrap();
         uniform.sample(rng)
     }
 
@@ -296,19 +315,23 @@ pub trait UniformSampler: Sized {
         where B1: SampleBorrow<Self::X> + Sized,
               B2: SampleBorrow<Self::X> + Sized
     {
-        let uniform: Self = UniformSampler::new_inclusive(low, high);
+        let uniform: Self = UniformSampler::new_inclusive(low, high).unwrap();
         uniform.sample(rng)
     }
 }
 
-impl<X: SampleUniform> From<Range<X>> for Uniform<X> {
-    fn from(r: ::core::ops::Range<X>) -> Uniform<X> {
+impl<X: SampleUniform> TryFrom<Range<X>> for Uniform<X> {
+    type Error = Error;
+
+    fn try_from(r: ::core::ops::Range<X>) -> Result<Uniform<X>, Error> {
         Uniform::new(r.start, r.end)
     }
 }
 
-impl<X: SampleUniform> From<RangeInclusive<X>> for Uniform<X> {
-    fn from(r: ::core::ops::RangeInclusive<X>) -> Uniform<X> {
+impl<X: SampleUniform> TryFrom<RangeInclusive<X>> for Uniform<X> {
+    type Error = Error;
+
+    fn try_from(r: ::core::ops::RangeInclusive<X>) -> Result<Uniform<X>, Error> {
         Uniform::new_inclusive(r.start(), r.end())
     }
 }
@@ -442,30 +465,31 @@ macro_rules! uniform_int_impl {
 
             #[inline] // if the range is constant, this helps LLVM to do the
                       // calculations at compile-time.
-            fn new<B1, B2>(low_b: B1, high_b: B2) -> Self
+            fn new<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
             where
                 B1: SampleBorrow<Self::X> + Sized,
                 B2: SampleBorrow<Self::X> + Sized,
             {
                 let low = *low_b.borrow();
                 let high = *high_b.borrow();
-                assert!(low < high, "Uniform::new called with `low >= high`");
+                if !(low < high) {
+                    return Err(Error::EmptyRange);
+                }
                 UniformSampler::new_inclusive(low, high - 1)
             }
 
             #[inline] // if the range is constant, this helps LLVM to do the
                       // calculations at compile-time.
-            fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Self
+            fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
             where
                 B1: SampleBorrow<Self::X> + Sized,
                 B2: SampleBorrow<Self::X> + Sized,
             {
                 let low = *low_b.borrow();
                 let high = *high_b.borrow();
-                assert!(
-                    low <= high,
-                    "Uniform::new_inclusive called with `low > high`"
-                );
+                if !(low <= high) {
+                    return Err(Error::EmptyRange);
+                }
                 let unsigned_max = ::core::$u_large::MAX;
 
                 let range = high.wrapping_sub(low).wrapping_add(1) as $unsigned;
@@ -476,12 +500,12 @@ macro_rules! uniform_int_impl {
                     0
                 };
 
-                UniformInt {
+                Ok(UniformInt {
                     low,
                     // These are really $unsigned values, but store as $ty:
                     range: range as $ty,
                     z: ints_to_reject as $unsigned as $ty,
-                }
+                })
             }
 
             #[inline]
@@ -589,26 +613,29 @@ macro_rules! uniform_simd_int_impl {
 
             #[inline] // if the range is constant, this helps LLVM to do the
                       // calculations at compile-time.
-            fn new<B1, B2>(low_b: B1, high_b: B2) -> Self
+            fn new<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
                 where B1: SampleBorrow<Self::X> + Sized,
                       B2: SampleBorrow<Self::X> + Sized
             {
                 let low = *low_b.borrow();
                 let high = *high_b.borrow();
-                assert!(low.lt(high).all(), "Uniform::new called with `low >= high`");
+                if !(low.lt(high).all()) {
+                    return Err(Error::EmptyRange);
+                }
                 UniformSampler::new_inclusive(low, high - 1)
             }
 
             #[inline] // if the range is constant, this helps LLVM to do the
                       // calculations at compile-time.
-            fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Self
+            fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
                 where B1: SampleBorrow<Self::X> + Sized,
                       B2: SampleBorrow<Self::X> + Sized
             {
                 let low = *low_b.borrow();
                 let high = *high_b.borrow();
-                assert!(low.le(high).all(),
-                        "Uniform::new_inclusive called with `low > high`");
+                if !(low.le(high).all()) {
+                    return Err(Error::EmptyRange);
+                }
                 let unsigned_max = ::core::$u_scalar::MAX;
 
                 // NOTE: these may need to be replaced with explicitly
@@ -625,12 +652,12 @@ macro_rules! uniform_simd_int_impl {
                 // zero which means only one sample is needed.
                 let zone = unsigned_max - ints_to_reject;
 
-                UniformInt {
+                Ok(UniformInt {
                     low,
                     // These are really $unsigned values, but store as $ty:
                     range: range.cast(),
                     z: zone.cast(),
-                }
+                })
             }
 
             fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Self::X {
@@ -751,7 +778,7 @@ impl UniformSampler for UniformChar {
 
     #[inline] // if the range is constant, this helps LLVM to do the
               // calculations at compile-time.
-    fn new<B1, B2>(low_b: B1, high_b: B2) -> Self
+    fn new<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
     where
         B1: SampleBorrow<Self::X> + Sized,
         B2: SampleBorrow<Self::X> + Sized,
@@ -759,12 +786,12 @@ impl UniformSampler for UniformChar {
         let low = char_to_comp_u32(*low_b.borrow());
         let high = char_to_comp_u32(*high_b.borrow());
         let sampler = UniformInt::<u32>::new(low, high);
-        UniformChar { sampler }
+        sampler.map(|sampler| UniformChar { sampler })
     }
 
     #[inline] // if the range is constant, this helps LLVM to do the
               // calculations at compile-time.
-    fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Self
+    fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
     where
         B1: SampleBorrow<Self::X> + Sized,
         B2: SampleBorrow<Self::X> + Sized,
@@ -772,7 +799,7 @@ impl UniformSampler for UniformChar {
         let low = char_to_comp_u32(*low_b.borrow());
         let high = char_to_comp_u32(*high_b.borrow());
         let sampler = UniformInt::<u32>::new_inclusive(low, high);
-        UniformChar { sampler }
+        sampler.map(|sampler| UniformChar { sampler })
     }
 
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Self::X {
@@ -822,28 +849,29 @@ macro_rules! uniform_float_impl {
         impl UniformSampler for UniformFloat<$ty> {
             type X = $ty;
 
-            fn new<B1, B2>(low_b: B1, high_b: B2) -> Self
+            fn new<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
             where
                 B1: SampleBorrow<Self::X> + Sized,
                 B2: SampleBorrow<Self::X> + Sized,
             {
                 let low = *low_b.borrow();
                 let high = *high_b.borrow();
-                debug_assert!(
-                    low.all_finite(),
-                    "Uniform::new called with `low` non-finite."
-                );
-                debug_assert!(
-                    high.all_finite(),
-                    "Uniform::new called with `high` non-finite."
-                );
-                assert!(low.all_lt(high), "Uniform::new called with `low >= high`");
+                /*
+                if !(low.all_finite()) || !(high.all_finite()) {
+                    return Err(Error::NonFinite);
+                }
+                */
+                if !(low.all_lt(high)) {
+                    return Err(Error::EmptyRange);
+                }
                 let max_rand = <$ty>::splat(
                     (::core::$u_scalar::MAX >> $bits_to_discard).into_float_with_exponent(0) - 1.0,
                 );
 
                 let mut scale = high - low;
-                assert!(scale.all_finite(), "Uniform::new: range overflow");
+                if !(scale.all_finite()) {
+                    return Err(Error::NonFinite);
+                }
 
                 loop {
                     let mask = (scale * max_rand + low).ge_mask(high);
@@ -855,34 +883,32 @@ macro_rules! uniform_float_impl {
 
                 debug_assert!(<$ty>::splat(0.0).all_le(scale));
 
-                UniformFloat { low, scale }
+                Ok(UniformFloat { low, scale })
             }
 
-            fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Self
+            fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
             where
                 B1: SampleBorrow<Self::X> + Sized,
                 B2: SampleBorrow<Self::X> + Sized,
             {
                 let low = *low_b.borrow();
                 let high = *high_b.borrow();
-                debug_assert!(
-                    low.all_finite(),
-                    "Uniform::new_inclusive called with `low` non-finite."
-                );
-                debug_assert!(
-                    high.all_finite(),
-                    "Uniform::new_inclusive called with `high` non-finite."
-                );
-                assert!(
-                    low.all_le(high),
-                    "Uniform::new_inclusive called with `low > high`"
-                );
+                /*
+                if !(low.all_finite()) || !(high.all_finite()) {
+                    return Err(Error::NonFinite);
+                }
+                */
+                if !low.all_le(high) {
+                    return Err(Error::EmptyRange);
+                }
                 let max_rand = <$ty>::splat(
                     (::core::$u_scalar::MAX >> $bits_to_discard).into_float_with_exponent(0) - 1.0,
                 );
 
                 let mut scale = (high - low) / max_rand;
-                assert!(scale.all_finite(), "Uniform::new_inclusive: range overflow");
+                if !scale.all_finite() {
+                    return Err(Error::NonFinite);
+                }
 
                 loop {
                     let mask = (scale * max_rand + low).gt_mask(high);
@@ -894,15 +920,15 @@ macro_rules! uniform_float_impl {
 
                 debug_assert!(<$ty>::splat(0.0).all_le(scale));
 
-                UniformFloat { low, scale }
+                Ok(UniformFloat { low, scale })
             }
 
             fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Self::X {
                 // Generate a value in the range [1, 2)
                 let value1_2 = (rng.gen::<$uty>() >> $bits_to_discard).into_float_with_exponent(0);
 
-                // Get a value in the range [0, 1) in order to avoid
-                // overflowing into infinity when multiplying with scale
+                // Get a value in the range [0, 1) to avoid overflowing into infinity when
+                // multiplying with scale.
                 let value0_1 = value1_2 - 1.0;
 
                 // We don't use `f64::mul_add`, because it is not available with
@@ -941,8 +967,8 @@ macro_rules! uniform_float_impl {
                     let value1_2 =
                         (rng.gen::<$uty>() >> $bits_to_discard).into_float_with_exponent(0);
 
-                    // Get a value in the range [0, 1) in order to avoid
-                    // overflowing into infinity when multiplying with scale
+                    // Get a value in the range [0, 1) to avoid overflowing into infinity when
+                    // multiplying with scale.
                     let value0_1 = value1_2 - 1.0;
 
                     // Doing multiply before addition allows some architectures
@@ -979,7 +1005,7 @@ macro_rules! uniform_float_impl {
                     // `high` are non-finite we'll end up here and can do the
                     // appropriate checks.
                     //
-                    // Likewise `high - low` overflowing to infinity is also
+                    // Likewise, `high - low` overflowing to infinity is also
                     // rare, so handle it here after the common case.
                     let mask = !scale.finite_mask();
                     if mask.any() {
@@ -1051,29 +1077,30 @@ impl UniformSampler for UniformDuration {
     type X = Duration;
 
     #[inline]
-    fn new<B1, B2>(low_b: B1, high_b: B2) -> Self
+    fn new<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
     where
         B1: SampleBorrow<Self::X> + Sized,
         B2: SampleBorrow<Self::X> + Sized,
     {
         let low = *low_b.borrow();
         let high = *high_b.borrow();
-        assert!(low < high, "Uniform::new called with `low >= high`");
+        if !(low < high) {
+            return Err(Error::EmptyRange);
+        }
         UniformDuration::new_inclusive(low, high - Duration::new(0, 1))
     }
 
     #[inline]
-    fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Self
+    fn new_inclusive<B1, B2>(low_b: B1, high_b: B2) -> Result<Self, Error>
     where
         B1: SampleBorrow<Self::X> + Sized,
         B2: SampleBorrow<Self::X> + Sized,
     {
         let low = *low_b.borrow();
         let high = *high_b.borrow();
-        assert!(
-            low <= high,
-            "Uniform::new_inclusive called with `low > high`"
-        );
+        if !(low <= high) {
+            return Err(Error::EmptyRange);
+        }
 
         let low_s = low.as_secs();
         let low_n = low.subsec_nanos();
@@ -1088,7 +1115,7 @@ impl UniformSampler for UniformDuration {
         let mode = if low_s == high_s {
             UniformDurationMode::Small {
                 secs: low_s,
-                nanos: Uniform::new_inclusive(low_n, high_n),
+                nanos: Uniform::new_inclusive(low_n, high_n)?,
             }
         } else {
             let max = high_s
@@ -1098,7 +1125,7 @@ impl UniformSampler for UniformDuration {
             if let Some(higher_bound) = max {
                 let lower_bound = low_s * 1_000_000_000 + u64::from(low_n);
                 UniformDurationMode::Medium {
-                    nanos: Uniform::new_inclusive(lower_bound, higher_bound),
+                    nanos: Uniform::new_inclusive(lower_bound, higher_bound)?,
                 }
             } else {
                 // An offset is applied to simplify generation of nanoseconds
@@ -1106,14 +1133,14 @@ impl UniformSampler for UniformDuration {
                 UniformDurationMode::Large {
                     max_secs: high_s,
                     max_nanos,
-                    secs: Uniform::new_inclusive(low_s, high_s),
+                    secs: Uniform::new_inclusive(low_s, high_s)?,
                 }
             }
         };
-        UniformDuration {
+        Ok(UniformDuration {
             mode,
             offset: low_n,
-        }
+        })
     }
 
     #[inline]
@@ -1133,7 +1160,7 @@ impl UniformSampler for UniformDuration {
                 secs,
             } => {
                 // constant folding means this is at least as fast as `Rng::sample(Range)`
-                let nano_range = Uniform::new(0, 1_000_000_000);
+                let nano_range = Uniform::new(0, 1_000_000_000).unwrap();
                 loop {
                     let s = secs.sample(rng);
                     let n = nano_range.sample(rng);
@@ -1155,7 +1182,7 @@ mod tests {
     #[test]
     #[cfg(feature = "serde1")]
     fn test_serialization_uniform_duration() {
-        let distr = UniformDuration::new(Duration::from_secs(10), Duration::from_secs(60));
+        let distr = UniformDuration::new(Duration::from_secs(10), Duration::from_secs(60)).unwrap();
         let de_distr: UniformDuration = bincode::deserialize(&bincode::serialize(&distr).unwrap()).unwrap();
         assert_eq!(
             distr.offset, de_distr.offset
@@ -1188,39 +1215,37 @@ mod tests {
     #[test]
     #[cfg(feature = "serde1")]
     fn test_uniform_serialization() {
-        let unit_box: Uniform<i32>  = Uniform::new(-1, 1);
+        let unit_box: Uniform<i32>  = Uniform::new(-1, 1).unwrap();
         let de_unit_box: Uniform<i32> = bincode::deserialize(&bincode::serialize(&unit_box).unwrap()).unwrap();
 
         assert_eq!(unit_box.0.low, de_unit_box.0.low);
         assert_eq!(unit_box.0.range, de_unit_box.0.range);
         assert_eq!(unit_box.0.z, de_unit_box.0.z);
 
-        let unit_box: Uniform<f32> = Uniform::new(-1., 1.);
+        let unit_box: Uniform<f32> = Uniform::new(-1., 1.).unwrap();
         let de_unit_box: Uniform<f32> = bincode::deserialize(&bincode::serialize(&unit_box).unwrap()).unwrap();
 
         assert_eq!(unit_box.0.low, de_unit_box.0.low);
         assert_eq!(unit_box.0.scale, de_unit_box.0.scale);
     }
 
-    #[should_panic]
     #[test]
     fn test_uniform_bad_limits_equal_int() {
-        Uniform::new(10, 10);
+        assert_eq!(Uniform::new(10, 10), Err(Error::EmptyRange));
     }
 
     #[test]
     fn test_uniform_good_limits_equal_int() {
         let mut rng = crate::test::rng(804);
-        let dist = Uniform::new_inclusive(10, 10);
+        let dist = Uniform::new_inclusive(10, 10).unwrap();
         for _ in 0..20 {
             assert_eq!(rng.sample(dist), 10);
         }
     }
 
-    #[should_panic]
     #[test]
     fn test_uniform_bad_limits_flipped_int() {
-        Uniform::new(10, 5);
+        assert_eq!(Uniform::new(10, 5), Err(Error::EmptyRange));
     }
 
     #[test]
@@ -1234,25 +1259,25 @@ mod tests {
         macro_rules! t {
             ($ty:ident, $v:expr, $le:expr, $lt:expr) => {{
                 for &(low, high) in $v.iter() {
-                    let my_uniform = Uniform::new(low, high);
+                    let my_uniform = Uniform::new(low, high).unwrap();
                     for _ in 0..1000 {
                         let v: $ty = rng.sample(my_uniform);
                         assert!($le(low, v) && $lt(v, high));
                     }
 
-                    let my_uniform = Uniform::new_inclusive(low, high);
+                    let my_uniform = Uniform::new_inclusive(low, high).unwrap();
                     for _ in 0..1000 {
                         let v: $ty = rng.sample(my_uniform);
                         assert!($le(low, v) && $le(v, high));
                     }
 
-                    let my_uniform = Uniform::new(&low, high);
+                    let my_uniform = Uniform::new(&low, high).unwrap();
                     for _ in 0..1000 {
                         let v: $ty = rng.sample(my_uniform);
                         assert!($le(low, v) && $lt(v, high));
                     }
 
-                    let my_uniform = Uniform::new_inclusive(&low, &high);
+                    let my_uniform = Uniform::new_inclusive(&low, &high).unwrap();
                     for _ in 0..1000 {
                         let v: $ty = rng.sample(my_uniform);
                         assert!($le(low, v) && $le(v, high));
@@ -1323,7 +1348,7 @@ mod tests {
         let d = Uniform::new(
             core::char::from_u32(0xD7F0).unwrap(),
             core::char::from_u32(0xE010).unwrap(),
-        );
+        ).unwrap();
         for _ in 0..100 {
             let c = d.sample(&mut rng);
             assert!((c as u32) < 0xD800 || (c as u32) > 0xDFFF);
@@ -1354,8 +1379,8 @@ mod tests {
                     for lane in 0..<$ty>::lanes() {
                         let low = <$ty>::splat(0.0 as $f_scalar).replace(lane, low_scalar);
                         let high = <$ty>::splat(1.0 as $f_scalar).replace(lane, high_scalar);
-                        let my_uniform = Uniform::new(low, high);
-                        let my_incl_uniform = Uniform::new_inclusive(low, high);
+                        let my_uniform = Uniform::new(low, high).unwrap();
+                        let my_incl_uniform = Uniform::new_inclusive(low, high).unwrap();
                         for _ in 0..100 {
                             let v = rng.sample(my_uniform).extract(lane);
                             assert!(low_scalar <= v && v < high_scalar);
@@ -1367,7 +1392,7 @@ mod tests {
                         }
 
                         assert_eq!(
-                            rng.sample(Uniform::new_inclusive(low, low)).extract(lane),
+                            rng.sample(Uniform::new_inclusive(low, low).unwrap()).extract(lane),
                             low_scalar
                         );
 
@@ -1400,14 +1425,14 @@ mod tests {
                     rng.sample(Uniform::new_inclusive(
                         ::core::$f_scalar::MAX,
                         ::core::$f_scalar::MAX
-                    )),
+                    ).unwrap()),
                     ::core::$f_scalar::MAX
                 );
                 assert_eq!(
                     rng.sample(Uniform::new_inclusive(
                         -::core::$f_scalar::MAX,
                         -::core::$f_scalar::MAX
-                    )),
+                    ).unwrap()),
                     -::core::$f_scalar::MAX
                 );
             }};
@@ -1428,9 +1453,8 @@ mod tests {
     }
 
     #[test]
-    #[should_panic]
     fn test_float_overflow() {
-        let _ = Uniform::from(::core::f64::MIN..::core::f64::MAX);
+        assert_eq!(Uniform::try_from(::core::f64::MIN..::core::f64::MAX), Err(Error::NonFinite));
     }
 
     #[test]
@@ -1478,10 +1502,10 @@ mod tests {
                         let low = <$ty>::splat(0.0 as $f_scalar).replace(lane, low_scalar);
                         let high = <$ty>::splat(1.0 as $f_scalar).replace(lane, high_scalar);
                         assert!(catch_unwind(|| range(low, high)).is_err());
-                        assert!(catch_unwind(|| Uniform::new(low, high)).is_err());
-                        assert!(catch_unwind(|| Uniform::new_inclusive(low, high)).is_err());
+                        assert!(Uniform::new(low, high).is_err());
+                        assert!(Uniform::new_inclusive(low, high).is_err());
                         assert!(catch_unwind(|| range(low, low)).is_err());
-                        assert!(catch_unwind(|| Uniform::new(low, low)).is_err());
+                        assert!(Uniform::new(low, low).is_err());
                     }
                 }
             }};
@@ -1516,7 +1540,7 @@ mod tests {
             ),
         ];
         for &(low, high) in v.iter() {
-            let my_uniform = Uniform::new(low, high);
+            let my_uniform = Uniform::new(low, high).unwrap();
             for _ in 0..1000 {
                 let v = rng.sample(my_uniform);
                 assert!(low <= v && v < high);
@@ -1538,15 +1562,15 @@ mod tests {
         impl UniformSampler for UniformMyF32 {
             type X = MyF32;
 
-            fn new<B1, B2>(low: B1, high: B2) -> Self
+            fn new<B1, B2>(low: B1, high: B2) -> Result<Self, Error>
             where
                 B1: SampleBorrow<Self::X> + Sized,
                 B2: SampleBorrow<Self::X> + Sized,
             {
-                UniformMyF32(UniformFloat::<f32>::new(low.borrow().x, high.borrow().x))
+                UniformFloat::<f32>::new(low.borrow().x, high.borrow().x).map(UniformMyF32)
             }
 
-            fn new_inclusive<B1, B2>(low: B1, high: B2) -> Self
+            fn new_inclusive<B1, B2>(low: B1, high: B2) -> Result<Self, Error>
             where
                 B1: SampleBorrow<Self::X> + Sized,
                 B2: SampleBorrow<Self::X> + Sized,
@@ -1565,7 +1589,7 @@ mod tests {
         }
 
         let (low, high) = (MyF32 { x: 17.0f32 }, MyF32 { x: 22.0f32 });
-        let uniform = Uniform::new(low, high);
+        let uniform = Uniform::new(low, high).unwrap();
         let mut rng = crate::test::rng(804);
         for _ in 0..100 {
             let x: MyF32 = rng.sample(uniform);
@@ -1575,20 +1599,20 @@ mod tests {
 
     #[test]
     fn test_uniform_from_std_range() {
-        let r = Uniform::from(2u32..7);
+        let r = Uniform::try_from(2u32..7).unwrap();
         assert_eq!(r.0.low, 2);
         assert_eq!(r.0.range, 5);
-        let r = Uniform::from(2.0f64..7.0);
+        let r = Uniform::try_from(2.0f64..7.0).unwrap();
         assert_eq!(r.0.low, 2.0);
         assert_eq!(r.0.scale, 5.0);
     }
 
     #[test]
     fn test_uniform_from_std_range_inclusive() {
-        let r = Uniform::from(2u32..=6);
+        let r = Uniform::try_from(2u32..=6).unwrap();
         assert_eq!(r.0.low, 2);
         assert_eq!(r.0.range, 5);
-        let r = Uniform::from(2.0f64..=7.0);
+        let r = Uniform::try_from(2.0f64..=7.0).unwrap();
         assert_eq!(r.0.low, 2.0);
         assert!(r.0.scale > 5.0);
         assert!(r.0.scale < 5.0 + 1e-14);
@@ -1607,7 +1631,7 @@ mod tests {
             }
             assert_eq!(&buf, expected_single);
 
-            let distr = Uniform::new(lb, ub);
+            let distr = Uniform::new(lb, ub).unwrap();
             for x in &mut buf {
                 *x = rng.sample(&distr);
             }
@@ -1650,9 +1674,9 @@ mod tests {
 
     #[test]
     fn uniform_distributions_can_be_compared() {
-        assert_eq!(Uniform::new(1.0, 2.0), Uniform::new(1.0, 2.0));
+        assert_eq!(Uniform::new(1.0, 2.0).unwrap(), Uniform::new(1.0, 2.0).unwrap());
 
         // To cover UniformInt
-        assert_eq!(Uniform::new(1 as u32, 2 as u32), Uniform::new(1 as u32, 2 as u32));
+        assert_eq!(Uniform::new(1 as u32, 2 as u32).unwrap(), Uniform::new(1 as u32, 2 as u32).unwrap());
     }
 }